converted doubls to floats everywhere possible. NOTE still having blank-out problem (divide by zero hiding somewhere...)

+++ draw.h
@@ -8,9 +8,9 @@ class draw
 public:
 	draw();
 	~draw();
-	static void line(unsigned int, unsigned int, unsigned int, unsigned int, float, float, float);
+	static void line(float, float, float, float, float, float, float);
 	static void line3(float, float, float, float, float, float, float, float, float);
-	static void square(unsigned int, unsigned int, unsigned int, float, float, float);
+	static void square(float, float, float, float, float, float);
 	static void square3(float, float, float, float, float, float, float);
 	static void map_xy_xyz(int, int, float *);
 	static void mode2d(bool);
